arbitrary-fn-arity: remove hard arity cap, out-of-line per-arg storage (#669)
* Raise MAX_FN_ARITY to 64 with a soft lint above 16 (arbitrary-fn-arity Phase 1/6)

Lifts the historical hard cap of 16 positional parameters to 64 and softens
the failure mode past 16 to a lint nudge, executing the landable increment of
docs/upcoming/v2/arbitrary-fn-arity-plan.md.

Key move: shrink sizeof(Type) so the higher cap costs nothing. A naive bump of
the constant tripled Type (the fn union embeds inline per-arg arrays that every
Type carries) and overflowed codegen's deep by-value-Type recursion on nested
data/hamt literals. Instead:

  - Pack the eight inline bool arg_*[MAX_FN_ARITY] flag arrays into one
    uint8_t arg_flags[] byte per arg, behind FN_ARG_FLAG / FN_ARG_SET accessors.
  - Narrow arg_kinds from the 4-byte TypeKind enum to uint8_t (61 enumerators).

Net sizeof(Type) drops from 304 (old, cap 16) to 240 (new, cap 64) -- no
memory or stack regression despite the 4x cap.

Also:
  - Uncap the emit_fns.c param loops that were hard-coded i<16 to n_params;
    grow the fn-param struct arrays (ForcedAbiSpec/BorrowSpecInfo.arg_kinds,
    pbp_param_ptrs, needs_box_*) to MAX_FN_ARITY.
  - Track the cap in the interpreter (EVAL_MAX_FN_ARITY) and spice FFI
    (TUR_SPICE_MAX_ARITY); the FFI manifest parser already clean-rejects an
    over-cap export, so the raise is backward-compatible (no miscompile).
  - New TUR-W0041 warning: declaring >16 positional params points back at the
    Function Arity Style Guide instead of hard-erroring (fires at 17+, not 16).

Fixtures: 20-param call, 20-param partial application, 20-arg fn passed as a
value, and a 40-param fn -- all exercised in interpreter and compiled paths.
Full suite green (2115 passed, 0 failed). Truly unbounded (uint32_t) arity via
the out-of-line arg restructure + spice descriptor versioning remains as
documented follow-up.


* Remove the function-arity cap: unbounded positional parameters

Functions no longer have a hard parameter-count ceiling. A defn, fn, or
extern-c may declare an arbitrary number of positional parameters, bounded only
by uint32_t -- matching the emitted C, which has no limit of its own. The
previous change had raised the mechanical cap from 16 to 64; this removes it.

Core restructure (arbitrary-fn-arity-plan Phase 2):

  - Out-of-line per-arg storage. Type.as.fn.arg_kinds / arg_flags are arena
    pointers (from a process-lifetime global type arena, tur_fn_args_alloc)
    instead of inline [MAX_FN_ARITY] arrays; fn arity is uint32_t. sizeof(Type)
    drops from 304 to 128 -- the fn variant no longer dominates the union, so
    deep by-value-Type codegen recursion gains headroom. The few sites that
    rebuild a fn type at a different arity (dict/env injection, poly and
    struct-field instantiation, forward-decl growth) allocate fresh arrays
    rather than mutate a shared one.

  - FnDef.n_params / ExternC.n_params / EmitCtx.n_fn_params and the arity/param
    loop counters and count locals across elab, emit, and the passes were
    widened from uint8_t to uint32_t (a uint8_t FnDef.n_params was silently
    truncating 300 -> 44). The >= MAX_FN_ARITY hard-reject guards are gone.

  - Per-param buffers are arena/count-sized, not fixed [MAX_FN_ARITY]:
    elaboration param scratch (defn/fn/extern-c), fwd_decl_scan_params, the
    emit param/thunk/needs-box buffers, ABI-instantiation scratch, and pbp
    (pass-by-ptr) params, which is now a realloc-grown array on EmitCtx.

  - poly_arg_mask / poly_agg_arg_mask are uint64_t indexed via ARG_IDX_BIT(i)
    (0 for i >= 64) so ordinary high-arity calls never hit an undefined shift.

  - Fixed a latent bug the smaller Type union exposed: effect_check read
    .as.fn.effect_row off a higher-order param that may be TY_FORALL, not TY_FN,
    yielding a garbage pointer; now guarded on kind == TY_FN.

Exceeding the historical 16 params remains a TUR-W0041 lint nudge, not an error.
MAX_FN_ARITY (64) survives only as the default size of a few internal codegen
fast-path buffers (ABI specialization, interpreter effect-perform args), each of
which falls back gracefully / errors cleanly for wider functions.

Verified end to end (compiled) at 100/300/500/1000 ordinary params, an 80-param
generic function, and a 70-param struct-by-value function. New fixture
fn-arity-100-param. Full suite green (2116 passed, 0 failed).


* Make the spice-FFI export descriptor unbounded

The REPL spice loader's in-memory export descriptor no longer caps positional
arity at 64. Previously a single exported symbol with more than
TUR_SPICE_MAX_ARITY params made the whole exports.manifest parse fail
(return -1), taking every sibling export in the spice down with it.

  - TurSpiceExport.arg_classes is a heap array of length n_args (was a fixed
    char[TUR_SPICE_MAX_ARITY]); n_args is uint32_t. Ownership transfers from
    parse_manifest_line through append_export and is freed in spice_export_free.
  - parse_manifest_line grows its arg-class buffer on demand -- no cap guard.
  - load_manifest reads lines with getline instead of a fixed char[4096], so a
    very high-arity export's (multi-KB) manifest row is not truncated.
  - The FFI marshalling scratch (i_vals/f_vals) uses a TUR_SPICE_ARITY_FASTPATH
    (16) inline fast path with a heap spill above it, freed on every exit via a
    single cleanup path.

The manifest writer already emitted all n_params tags, and the manifest is
plain variable-length text, so no binary version gate is required -- an older
tur still cleanly rejects a wider row rather than misreading it.

The remaining interpreter-FFI limit is orthogonal to the descriptor: the
generated shape dispatcher (tur_ffi_thunk_call) covers arg-shapes only up to
gen_ffi_dispatch.py --max-arity (6 by default), so a REPL call whose shape
exceeds that fails at call time with a regenerate-the-table diagnostic. The
compiled path calls spice exports directly and has no such limit.

Verified end to end: a spice exporting a 100-param symbol now loads (2 exports),
its 1-arg sibling is callable (=> 12), the descriptor reports the true arity
(arity-mismatch error names 100 args), and a full 100-arg call fails cleanly
with the dispatcher diagnostic -- no crash, no whole-manifest rejection. Loader
allocations are leak-clean. Full suite green (2116 passed, 0 failed).


* Plan: arbitrary-arity spice calls in the interpreter (retire shape-dispatch ceiling)

Research + design doc for the last arity ceiling in the arbitrary-fn-arity
story: the interpreter/REPL path that calls a dlopen'ed spice export through a
generated shape-dispatch table baked into `tur` (exponential, 3 * 2^N shapes,
capped at --max-arity 6).

Documents why the current design is bounded and fixed at tur's build time,
honestly evaluates the currying angle (it organizes arg collection, which is
already unbounded; it does not solve the ABI placement of the single native
call), and recommends a spice-emitted per-export FFI shim -- generated during
`tur build --shared` where the exact signature is known, so it scales to any
arity/shape with no new dependency and no runtime codegen (precedent:
ensure_typed_fatshim). Covers libffi and manifest-mined-shape alternatives,
compatibility/fallback for pre-shim spices, phasing, tests, and non-goals.


---------

+++ src/compiler/emit_core.c
@@ -1617,7 +1617,7 @@ bool emit_reresolve_disp_type(EmitCtx *ctx, const Expr *call,
                 ft->as.fn.result_full_type->kind == TY_TYVAR &&
                 ft->as.fn.result_full_type->as.tyvar_.name) {
                 const char *rname = ft->as.fn.result_full_type->as.tyvar_.name;
-                uint8_t np = ft->as.fn.arity;
+                uint32_t np = ft->as.fn.arity;
                 for (uint8_t pi = 0;
                      pi < np && pi < recv->as.call_.n_args; pi++) {
                     const Type *pft = ft->as.fn.arg_full_types[pi];
@@ -2225,7 +2225,7 @@ char *name_for_binding(EmitCtx *ctx, const Binding *b) {
     }
     /* Check if this binding is a function parameter in the current context */
     if (ctx->fn_params) {
-        for (uint8_t i = 0; i < ctx->n_fn_params; i++) {
+        for (uint32_t i = 0; i < ctx->n_fn_params; i++) {
             if (ctx->fn_params[i] == b) {
                 return raw_name_for_binding(b);
             }
@@ -3351,7 +3351,7 @@ Binding **collect_handle_captures(const Expr *body, uint32_t *n_out) {
                         for (uint32_t j = 0; j < n_icaps; j++) {
                             Binding *ib = icaps[j];
                             bool inner_local = (icase->k_binding && ib == icase->k_binding);
-                            for (uint8_t p = 0; p < icase->n_params && !inner_local; p++)
+                            for (uint32_t p = 0; p < icase->n_params && !inner_local; p++)
                                 if (icase->param_bindings && icase->param_bindings[p] == ib)
                                     inner_local = true;
                             if (!inner_local)
